apps/api/pyproject.toml
[project]
name = "lms-api"
version = "0.1.0"
description = "Library Management System gRPC API"
requires-python = ">=3.14"
dependencies = [
  "grpcio==1.73.0",
  "grpcio-tools==1.73.0",
  "protobuf>=6.30.0,<7.0.0",
  "psycopg[binary]>=3.2.9",
]

[dependency-groups]
dev = [
  "pytest==8.3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["library_api"]

[tool.pytest.ini_options]
pythonpath = [".", "../packages/proto/gen/python"]
markers = [
  "integration: integration tests requiring postgres",
]
